restructure project according to ddd patern
