--- patientcare/src/main/java/com/medieninformatik/patientcare/api/HelloController.java
@@ -1,28 +1,0 @@
-package com.medieninformatik.patientcare.api;
-
-import com.medieninformatik.patientcare.services.HelloService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-
-
-
-@RestController
-@RequestMapping(path = "/hello")
-public class HelloController {
-
-    private final HelloService helloService;
-
-    @Autowired
-    public HelloController(HelloService helloService) {
-        this.helloService = helloService;
-    }
-
-    @GetMapping(path = "/sayhello")
-    public String gethello() {
-        return helloService.sayHello();
-    }
-}
