--- patientcare/src/main/java/com/medieninformatik/patientcare/repo/AppointmentRepo.java
@@ -1,10 +1,0 @@
-package com.medieninformatik.patientcare.repo;
-
-import com.medieninformatik.patientcare.entities.Appointment;
-import com.medieninformatik.patientcare.entities.Doctor;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AppointmentRepo extends JpaRepository<Appointment, Long>{
-}
