+++ patientcare/src/main/java/com/medieninformatik/patientcare/userManagement/domain/model/Doctor.java
@@ -1,0 +1,28 @@
+package com.medieninformatik.patientcare.userManagement.domain.model;
+
+import com.medieninformatik.patientcare.userManagement.domain.model.shared.User;
+import jakarta.persistence.*;
+
+@Entity
+public class Doctor
+        extends User {
+
+    // Default constructor
+    public Doctor() {
+    }
+
+    // Constructor with parameters
+    public Doctor(String firstName, String lastName) {
+        this.setFirstName(firstName);
+        this.setLastName(lastName);
+    }
+
+    @Override
+    public String toString() {
+        return "Doctor{" +
+                "id=" + super.getId() +
+                ", firstName='" + super.getFirstName() + '\'' +
+                ", lastName='" + super.getLastName() + '\'' +
+                '}';
+    }
+}
