+++ patientcare/src/main/java/com/medieninformatik/patientcare/userManagement/services/PatientService.java
@@ -1,0 +1,23 @@
+package com.medieninformatik.patientcare.userManagement.services;
+
+import com.medieninformatik.patientcare.userManagement.domain.model.Patient;
+import com.medieninformatik.patientcare.userManagement.infrastructure.repositories.PatientRepo;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+@Service
+public class PatientService {
+
+    private final PatientRepo patientRepo;
+
+    @Autowired
+    public PatientService(PatientRepo patientRepo) {
+        this.patientRepo = patientRepo;
+    }
+
+    public Optional<Patient> getPatient(Long personId) {
+        return patientRepo.findById(personId);
+    }
+}
